Resolve cyclic dependencies (4 files)
- `src/components/dialogs/parameters/network-parameters.tsx`
- `src/components/dialogs/parameters/parameters.jsx`
- `src/components/dialogs/parameters/common/computation-parameters-util.ts`
- `src/redux/reducer.ts`
