+++ src/components/dialogs/parameters/dynamicsimulation/solver-parameters.jsx
@@ -14,7 +14,7 @@ import IdaSolverParameters, { getFormSchema as getIdaFormSchema } from './solver
 import SimplifiedSolverParameters, {
     getFormSchema as getSimplifiedFormSchema,
 } from './solver/simplified-solver-parameters';
-import { TabPanel } from '../parameters';
+import TabPanel from '../widget/tab-panel';
 
 const SOLVER_TYPES = {
     IDA: 'IDA',
