+++ src/components/dialogs/parameters/flux-conventions.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright © 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+export enum FluxConventions {
+    IIDM = 'iidm',
+    TARGET = 'target',
+}
