+++ src/components/dialogs/parameters/hooks.jsx
@@ -5,216 +5,12 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import React, { useCallback, useEffect, useState, useRef, useMemo } from 'react';
-import { FormattedMessage } from 'react-intl';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { Grid, Box, Button, Typography, Switch, Select, MenuItem } from '@mui/material';
-
-import { useSnackMessage, useDebounce } from '@gridsuite/commons-ui';
+import { useDebounce, useSnackMessage } from '@gridsuite/commons-ui';
 import { OptionalServicesStatus } from 'components/utils/optional-services';
 import { updateConfigParameter } from 'services/config';
 import { isComputationParametersUpdated } from './common/computation-parameters-util';
-
-export const CloseButton = ({ hideParameters, ...props }) => {
-    return <LabelledButton callback={hideParameters} label={'close'} {...props} />;
-};
-
-export const LabelledButton = ({ callback, label, ...props }) => {
-    return (
-        <Button onClick={callback} {...props}>
-            <FormattedMessage id={label} />
-        </Button>
-    );
-};
-
-export const SwitchWithLabel = ({ value, label, callback }) => {
-    return (
-        <>
-            <Grid item xs={8} sx={styles.parameterName}>
-                <FormattedMessage id={label} />
-            </Grid>
-            <Grid item container xs={4} sx={styles.controlItem}>
-                <Switch
-                    checked={value}
-                    onChange={callback}
-                    value={value}
-                    inputProps={{ 'aria-label': 'primary checkbox' }}
-                />
-            </Grid>
-        </>
-    );
-};
-
-export const DropDown = ({ value, label, values, callback }) => {
-    return (
-        <>
-            <Grid item xs={5} sx={styles.parameterName}>
-                <FormattedMessage id={label} />
-            </Grid>
-            <Grid item container xs={4} sx={styles.controlItem}>
-                <Select labelId={label} value={value} onChange={callback} size="small">
-                    {Object.entries(values).map(([key, value]) => (
-                        <MenuItem key={key} value={key}>
-                            <FormattedMessage id={value} />
-                        </MenuItem>
-                    ))}
-                </Select>
-            </Grid>
-        </>
-    );
-};
-
-export const styles = {
-    title: (theme) => ({
-        padding: theme.spacing(2),
-    }),
-    minWidthMedium: (theme) => ({
-        minWidth: theme.spacing(20),
-    }),
-    parameterName: (theme) => ({
-        fontWeight: 'bold',
-        marginTop: theme.spacing(1),
-    }),
-    controlItem: {
-        justifyContent: 'flex-end',
-        flexGrow: 1,
-    },
-    controlParametersItem: {
-        justifyContent: 'flex-start',
-        flexGrow: 1,
-        height: 'fit-content',
-        paddingBottom: 4,
-    },
-    button: (theme) => ({
-        marginBottom: theme.spacing(2),
-        marginLeft: theme.spacing(1),
-    }),
-    subgroupParameters: (theme) => ({
-        marginTop: theme.spacing(2),
-        marginBottom: theme.spacing(1),
-    }),
-    subgroupParametersAccordion: {
-        '&:before': {
-            display: 'none',
-        },
-        background: 'none',
-    },
-    subgroupParametersAccordionSummary: (theme) => ({
-        flexDirection: 'row-reverse',
-        '& .MuiAccordionSummary-expandIconWrapper': {
-            transform: 'rotate(-90deg)',
-        },
-        '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
-            transform: 'rotate(0deg)',
-        },
-        '& .MuiAccordionSummary-content': {
-            marginLeft: theme.spacing(0),
-        },
-    }),
-    subgroupParametersAccordionDetails: (theme) => ({
-        padding: theme.spacing(0),
-    }),
-    marginTopButton: {
-        marginTop: '10px',
-        position: 'sticky',
-        bottom: 0,
-    },
-    scrollableGrid: (theme) => ({
-        overflowY: 'auto',
-        overflowX: 'hidden',
-        maxHeight: '85%', // TODO This needs to be refactored
-        paddingRight: theme.spacing(2),
-        paddingTop: theme.spacing(2),
-        paddingBottom: theme.spacing(1),
-        flexGrow: 1,
-    }),
-    singleItem: (theme) => ({
-        display: 'flex',
-        flex: 'auto',
-        alignItems: 'flex-start',
-        justifyContent: 'space-between',
-        marginTop: theme.spacing(1),
-        marginBottom: theme.spacing(1),
-    }),
-    firstTextField: (theme) => ({
-        marginLeft: theme.spacing(3),
-    }),
-    secondTextField: (theme) => ({
-        marginLeft: theme.spacing(3),
-        marginRight: theme.spacing(2),
-    }),
-    singleTextField: (theme) => ({
-        display: 'flex',
-        marginRight: theme.spacing(2),
-        marginLeft: theme.spacing(1),
-    }),
-    tooltip: (theme) => ({
-        marginLeft: theme.spacing(1),
-    }),
-    text: (theme) => ({
-        display: 'flex',
-        marginBottom: theme.spacing(1),
-        marginTop: theme.spacing(1),
-    }),
-    multipleItems: (theme) => ({
-        display: 'flex',
-        flex: 'auto',
-        alignItems: 'flex-start',
-        justifyContent: 'space-between',
-        marginTop: theme.spacing(1),
-        marginBottom: theme.spacing(1),
-    }),
-    tabWithError: (theme) => ({
-        '&.Mui-selected': { color: theme.palette.error.main },
-        color: theme.palette.error.main,
-    }),
-    tabWithErrorIndicator: (theme) => ({
-        backgroundColor: theme.palette.error.main,
-    }),
-    panel: (theme) => ({
-        marginTop: theme.spacing(2),
-        marginBottom: theme.spacing(1),
-    }),
-    adjustExistingLimitsInfo: (theme) => ({
-        display: 'flex',
-        alignItems: 'center',
-        marginTop: theme.spacing(1),
-        marginBottom: theme.spacing(1),
-    }),
-    circularProgress: (theme) => ({
-        marginRight: theme.spacing(2),
-        color: theme.palette.primary.contrastText,
-    }),
-    icon: (theme) => ({
-        width: theme.spacing(3),
-    }),
-    modificationsTitle: (theme) => ({
-        display: 'flex',
-        alignItems: 'center',
-        margin: theme.spacing(0),
-        padding: theme.spacing(1),
-        backgroundColor: theme.palette.primary.main,
-        color: theme.palette.primary.contrastText,
-        overflow: 'hidden',
-    }),
-};
-
-export const TabPanel = (props) => {
-    const { children, value, index, keepState, ...other } = props;
-    return (
-        <Typography
-            component="div"
-            role="tabpanel"
-            hidden={value !== index}
-            id={`simple-tabpanel-${index}`}
-            aria-labelledby={`simple-tab-${index}`}
-            style={{ flexGrow: 1 }}
-            {...other}
-        >
-            {(value === index || keepState) && <Box sx={styles.panel}>{children}</Box>}
-        </Typography>
-    );
-};
 
 const INITIAL_PROVIDERS = {};
