+++ src/components/dialogs/parameters/network-visualizations/map-parameters.tsx
@@ -20,10 +20,11 @@ import {
     PARAM_MAP_MANUAL_REFRESH,
 } from '../../../../utils/config-params';
 import { LineSeparator } from '../../dialogUtils';
-import { styles, useParameterState } from '../parameters';
+import { useParameterState } from '../hooks';
 import ParameterLineDropdown from '../widget/parameter-line-dropdown';
 import ParameterLineSlider from '../widget/parameter-line-slider';
 import ParameterLineSwitch from '../widget/parameter-line-switch';
+import styles from '../styles';
 
 export const MapParameters = () => {
     const alertThresholdMarks = [
