+++ src/components/dialogs/parameters/network-visualizations/network-area-diagram-parameters.tsx
@@ -6,8 +6,8 @@
  */
 import { Grid } from '@mui/material';
 import { PARAM_INIT_NAD_WITH_GEO_DATA } from '../../../../utils/config-params';
-import { styles } from '../parameters';
 import ParameterLineSwitch from '../widget/parameter-line-switch';
+import styles from '../styles';
 
 export const NetworkAreaDiagramParameters = () => {
     return (
