+++ src/components/dialogs/parameters/network-visualizations/single-line-diagram-parameters.jsx
@@ -16,10 +16,10 @@ import {
     PARAM_SUBSTATION_LAYOUT,
 } from '../../../../utils/config-params';
 import { LineSeparator } from '../../dialogUtils';
-import { styles } from '../parameters';
 import { SubstationLayout } from '../../../diagrams/diagram-common';
 import ParameterLineDropdown from '../widget/parameter-line-dropdown';
 import ParameterLineSwitch from '../widget/parameter-line-switch';
+import styles from '../styles';
 
 export const useGetAvailableComponentLibraries = (user) => {
     const [componentLibraries, setComponentLibraries] = useState([]);
