+++ src/components/dialogs/parameters/security-analysis/security-analysis-parameters-selector.tsx
@@ -11,11 +11,12 @@ import { FormattedMessage } from 'react-intl';
 
 import Grid from '@mui/material/Grid';
 import { Tab, Tabs } from '@mui/material';
-import { TabPanel, useParameterState } from '../parameters';
+import { useParameterState } from '../hooks';
 import { TAB_INFO, TAB_VALUES } from '../common/limitreductions/columns-definitions';
 import ViolationsHidingParameters from './security-analysis-violations-hiding';
 import LimitReductionsTableForm from '../common/limitreductions/limit-reductions-table-form';
 import { PARAM_DEVELOPER_MODE } from '../../../../utils/config-params';
+import TabPanel from '../widget/tab-panel';
 
 const SecurityAnalysisParametersSelector: FunctionComponent<{
     params: Record<string, any>;
