+++ src/components/dialogs/parameters/widget/index.ts
@@ -5,6 +5,17 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+export { default as LabelledButton } from './labelled-button';
+export type * from './labelled-button';
+export { default as CloseButton } from './close-button';
+export type * from './close-button';
+export { default as DropDown } from './drop-down';
+export type * from './drop-down';
+export { default as SwitchWithLabel } from './switch-with-label';
+export type * from './switch-with-label';
+export { default as TabPanel } from './tab-panel';
+export type * from './tab-panel';
+
 export * from './parameter-float';
 export * from './parameter-group';
 export { default as parameterLineDirectoryItemsInput } from './parameter-line-directory-items-input';
