+++ src/components/dialogs/parameters/widget/parameter-group.tsx
@@ -6,10 +6,45 @@
  */
 
 import { ExpandCircleDown, ExpandMore, Settings as SettingsIcon } from '@mui/icons-material';
-import { Accordion, AccordionDetails, AccordionSummary, Grid, Typography } from '@mui/material';
+import {
+    Accordion,
+    AccordionDetails,
+    AccordionSummary,
+    Grid,
+    type SxProps,
+    type Theme,
+    Typography,
+} from '@mui/material';
 import { FunctionComponent, PropsWithChildren, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
-import { styles } from '../parameters';
+
+const styles = {
+    subgroupParameters: (theme) => ({
+        marginTop: theme.spacing(2),
+        marginBottom: theme.spacing(1),
+    }),
+    subgroupParametersAccordion: {
+        '&:before': {
+            display: 'none',
+        },
+        background: 'none',
+    },
+    subgroupParametersAccordionSummary: (theme) => ({
+        flexDirection: 'row-reverse',
+        '& .MuiAccordionSummary-expandIconWrapper': {
+            transform: 'rotate(-90deg)',
+        },
+        '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
+            transform: 'rotate(0deg)',
+        },
+        '& .MuiAccordionSummary-content': {
+            marginLeft: theme.spacing(0),
+        },
+    }),
+    subgroupParametersAccordionDetails: (theme) => ({
+        padding: theme.spacing(0),
+    }),
+} as const satisfies Record<string, SxProps<Theme>>;
 
 type ParameterGroupProps = {
     label: string;
@@ -24,44 +59,42 @@ export const ParameterGroup: FunctionComponent<PropsWithChildren<ParameterGroupP
     const [mouseHover, setMouseHover] = useState(false);
 
     return (
-        <>
-            <Grid item xs={12} sx={styles.subgroupParameters}>
-                <Accordion
-                    sx={styles.subgroupParametersAccordion}
-                    expanded={props.state}
-                    onChange={(event, showed) => props.onClick(showed)}
-                    disableGutters
-                    elevation={0}
-                    square
-                    TransitionProps={{
-                        unmountOnExit: props.unmountOnExit || false,
-                    }}
-                    disabled={props.disabled || undefined}
+        <Grid item xs={12} sx={styles.subgroupParameters}>
+            <Accordion
+                sx={styles.subgroupParametersAccordion}
+                expanded={props.state}
+                onChange={(event, showed) => props.onClick(showed)}
+                disableGutters
+                elevation={0}
+                square
+                TransitionProps={{
+                    unmountOnExit: props.unmountOnExit || false,
+                }}
+                disabled={props.disabled || undefined}
+            >
+                <AccordionSummary
+                    sx={styles.subgroupParametersAccordionSummary}
+                    expandIcon={mouseHover ? <ExpandCircleDown /> : <ExpandMore />}
+                    onMouseEnter={(event) => setMouseHover(true)}
+                    onMouseLeave={(event) => setMouseHover(false)}
                 >
-                    <AccordionSummary
-                        sx={styles.subgroupParametersAccordionSummary}
-                        expandIcon={mouseHover ? <ExpandCircleDown /> : <ExpandMore />}
-                        onMouseEnter={(event) => setMouseHover(true)}
-                        onMouseLeave={(event) => setMouseHover(false)}
-                    >
-                        <SettingsIcon />
-                        <Typography sx={{ width: '66%', flexShrink: 0 }}>
-                            <FormattedMessage id={props.label} />
+                    <SettingsIcon />
+                    <Typography sx={{ width: '66%', flexShrink: 0 }}>
+                        <FormattedMessage id={props.label} />
+                    </Typography>
+                    {props.infoText && (
+                        <Typography
+                            sx={{ color: 'text.secondary', width: '34%' }}
+                            noWrap={true}
+                            align="right"
+                            variant="body2"
+                        >
+                            {props.infoText}
                         </Typography>
-                        {props.infoText && (
-                            <Typography
-                                sx={{ color: 'text.secondary', width: '34%' }}
-                                noWrap={true}
-                                align="right"
-                                variant="body2"
-                            >
-                                {props.infoText}
-                            </Typography>
-                        )}
-                    </AccordionSummary>
-                    <AccordionDetails sx={styles.subgroupParametersAccordionDetails}>{props.children}</AccordionDetails>
-                </Accordion>
-            </Grid>
-        </>
+                    )}
+                </AccordionSummary>
+                <AccordionDetails sx={styles.subgroupParametersAccordionDetails}>{props.children}</AccordionDetails>
+            </Accordion>
+        </Grid>
     );
 };
