+++ src/components/menus/equipment-menu.jsx
@@ -7,7 +7,7 @@
 
 import React, { useCallback, useMemo } from 'react';
 import Menu from '@mui/material/Menu';
-import { useParameterState } from '../dialogs/parameters/parameters';
+import { useParameterState } from '../dialogs/parameters/hooks';
 import { PARAM_DEVELOPER_MODE } from '../../utils/config-params';
 import { getEventType } from '../dialogs/dynamicsimulation/event/model/event.model';
 import { useSelector } from 'react-redux';
