+++ src/components/parameters-tabs.tsx
@@ -10,7 +10,7 @@ import { FormattedMessage } from 'react-intl';
 import { useSelector } from 'react-redux';
 import { Box, darken, DialogContentText, Divider, Grid, lighten, Tab, Tabs, Typography } from '@mui/material';
 import { Theme } from '@mui/material/styles';
-import { useParametersBackend, useParameterState } from './dialogs/parameters/parameters';
+import { useParametersBackend, useParameterState } from './dialogs/parameters/hooks';
 import { PARAM_DEVELOPER_MODE } from 'utils/config-params';
 import { useOptionalServiceStatus } from 'hooks/use-optional-service-status';
 import { OptionalServicesNames, OptionalServicesStatus } from './utils/optional-services';
